+++ Showcase/Showcase/Pages/Utils/ShowcaseElementsPage.swift
@@ -45,7 +45,7 @@ struct ShowcaseElementsPage: View {
                 }
                 .padding(.all, theme.spaceFixedMedium)
                 .navigationbarMenuForThemeSelection()
-                .oudsNavigationTitle("app_bottomBar_tokens")
+                .oudsNavigationTitle("app_bottomBar_tokens_label")
             }
             .background(theme.colorBackgroundPrimary.color(for: colorScheme))
         }